Crate pallet_multisig

Multisig pallet

A pallet for doing multisig dispatch.

Config

Call

Overview

This pallet contains functionality for multi-signature dispatch, a (potentially) stateful operation, allowing multiple signed origins (accounts) to coordinate and dispatch a call from a well-known origin, derivable deterministically from the set of account IDs and the threshold number of accounts from the set that must approve it. In the case that the threshold is just one then this is a stateless operation. This is useful for multisig wallets where cryptographic threshold signatures are not available or desired.

Interface

Dispatchable Functions

as_multi - Approve and if possible dispatch a call from a composite origin formed from a number of signed origins.

approve_as_multi - Approve a call from a composite origin.

cancel_as_multi - Cancel a call from a composite origin.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_multisig

Macros

log

Structs

Multisig

An open multisig operation.

Timepoint

A global extrinsic index, formed as the extrinsic index within a block, together with that block’s height. This allows a transaction in which a multisig operation of a particular composite was created to be uniquely identified.

Constants

LOG_TARGET

The log target of this pallet.