Function pallet_multisig::pallet::dispatchables::approve_as_multi

pub fn approve_as_multi<T: Config>(
    threshold: u16,
    other_signatories: Vec<T::AccountId>,
    maybe_timepoint: Option<Timepoint<BlockNumberFor<T>>>,
    call_hash: [u8; 32],
    max_weight: Weight,
)

Register approval for a dispatch to be made from a deterministic composite account if approved by a total of threshold - 1 of other_signatories.

Payment: DepositBase will be reserved if this is the first approval, plus threshold times DepositFactor. It is returned once this dispatch happens or is cancelled.

The dispatch origin for this call must be Signed.

threshold: The total number of approvals for this dispatch before it is executed.

other_signatories: The accounts (other than the sender) who can approve this dispatch. May not be empty.

maybe_timepoint: If this is the first approval, then this must be None. If it is not the first approval, then it must be Some, with the timepoint (block number and transaction index) of the first approval transaction.

call_hash: The hash of the call to be executed.

NOTE: If this is the final approval, you will want to use as_multi instead.

Complexity

O(S).

Up to one balance-reserve or unreserve operation.

One passthrough operation, one insert, both O(S) where S is the number of signatories. S is capped by MaxSignatories, with weight being proportional.

One encode & hash, both of complexity O(S).

Up to one binary search and insert (O(logS + S)).

I/O: 1 read O(S), up to 1 mutate O(S). Up to one remove.

One event.

Storage: inserts one item, value size bounded by MaxSignatories, with a deposit taken for its lifetime of DepositBase + threshold * DepositFactor.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::approve_as_multi.