pallet_multisig::pallet::dispatchables

Function as_multi

pub fn as_multi<T: Config>(
    threshold: u16,
    other_signatories: Vec<T::AccountId>,
    maybe_timepoint: Option<Timepoint<BlockNumberFor<T>>>,
    call: Box<<T as Config>::RuntimeCall>,
    max_weight: Weight,
)

Register approval for a dispatch to be made from a deterministic composite account if approved by a total of threshold - 1 of other_signatories.

If the approval threshold is met (including the sender’s approval), this will immediately execute the call. This is the only way to execute a multisig call - approve_as_multi will never trigger execution.

Payment: DepositBase will be reserved if this is the first approval, plus threshold times DepositFactor. It is returned once this dispatch happens or is cancelled.

The dispatch origin for this call must be Signed.

threshold: The total number of approvals for this dispatch before it is executed.

other_signatories: The accounts (other than the sender) who can approve this dispatch. May not be empty.

maybe_timepoint: If this is the first approval, then this must be None. If it is not the first approval, then it must be Some, with the timepoint (block number and transaction index) of the first approval transaction.

call: The call to be executed.

NOTE: For intermediate approvals (not the final approval), you should generally use approve_as_multi instead, since it only requires a hash of the call and is more efficient.

Result is equivalent to the dispatched result if threshold is exactly 1. Otherwise on success, result is Ok and the result from the interior call, if it was executed, may be found in the deposited MultisigExecuted event.

Complexity

O(S + Z + Call).

Up to one balance-reserve or unreserve operation.

One passthrough operation, one insert, both O(S) where S is the number of signatories. S is capped by MaxSignatories, with weight being proportional.

One call encode & hash, both of complexity O(Z) where Z is tx-len.

One encode & hash, both of complexity O(S).

Up to one binary search and insert (O(logS + S)).

I/O: 1 read O(S), up to 1 mutate O(S). Up to one remove.

One event.

The weight of the call.

Storage: inserts one item, value size bounded by MaxSignatories, with a deposit taken for its lifetime of DepositBase + threshold * DepositFactor.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::as_multi.