Crate pallet_multisig

Multisig Module

A module for doing multisig dispatch.

multisig::Config

Call

Overview

This module contains functionality for multi-signature dispatch, a (potentially) stateful operation, allowing multiple signed origins (accounts) to coordinate and dispatch a call from a well-known origin, derivable deterministically from the set of account IDs and the threshold number of accounts from the set that must approve it. In the case that the threshold is just one then this is a stateless operation. This is useful for multisig wallets where cryptographic threshold signatures are not available or desired.

Interface

Dispatchable Functions

as_multi - Approve and if possible dispatch a call from a composite origin formed from a number of signed origins.

approve_as_multi - Approve a call from a composite origin.

cancel_as_multi - Cancel a call from a composite origin.

Re-exports

pub use weights::WeightInfo;

pub use $crate::traits::GetPalletVersion as _;

Modules

weights

Weights for pallet_multisig THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 2.0.0 DATE: 2020-10-27, STEPS: [50, ], REPEAT: 20, LOW RANGE: [], HIGH RANGE: [] EXECUTION: Some(Wasm), WASM-EXECUTION: Compiled, CHAIN: Some(“dev”), DB CACHE: 128

Structs

Calls

Module

Multisig

An open multisig operation.

Multisigs

The set of open multisig operations.

Timepoint

A global extrinsic index, formed as the extrinsic index within a block, together with that block’s height. This allows a transaction in which a multisig operation of a particular composite was created to be uniquely identified.

Enums

Call

Dispatchable calls.

Error

RawEvent

Events for this module.

Traits

Config

Configuration trait.

Type Definitions

Event

RawEvent specialized for the configuration Config

OpaqueCall

Just a bunch of bytes, but they should decode to a valid Call.