Function pallet_multisig::pallet::dispatchables::cancel_as_multi

pub fn cancel_as_multi<T: Config>(
    threshold: u16,
    other_signatories: Vec<T::AccountId>,
    timepoint: Timepoint<BlockNumberFor<T>>,
    call_hash: [u8; 32]
)

Cancel a pre-existing, on-going multisig transaction. Any deposit reserved previously for this operation will be unreserved on success.

The dispatch origin for this call must be Signed.

threshold: The total number of approvals for this dispatch before it is executed.

other_signatories: The accounts (other than the sender) who can approve this dispatch. May not be empty.

timepoint: The timepoint (block number and transaction index) of the first approval transaction for this dispatch.

call_hash: The hash of the call to be executed.

Complexity

O(S).

Up to one balance-reserve or unreserve operation.

One passthrough operation, one insert, both O(S) where S is the number of signatories. S is capped by MaxSignatories, with weight being proportional.

One encode & hash, both of complexity O(S).

One event.

I/O: 1 read O(S), one remove.

Storage: removes one item.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::cancel_as_multi.